Divide classes into different packages
