+++ src/main/java/duke/CommandType.java
@@ -1,3 +1,5 @@
+package duke;
+
 public enum CommandType {
     BYE, LIST,
     TODO, EVENT, DEADLINE,
