+++ src/main/java/duke/Ui.java
@@ -1,3 +1,9 @@
+package duke;
+
+import duke.exception.DukeException;
+import duke.task.Task;
+import duke.task.TaskList;
+
 import java.util.Scanner;
 
 public class Ui {
