+++ src/main/java/duke/exception/DukeException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public abstract class DukeException extends Exception {
     @Override
     public String toString() {
